pyproject.toml
[tool.poetry]
name = "airbnb-220903"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "airbnb_220903"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
